Move dialets/postgres into schema/postgres test
